Rename Furnishing and implements tile for caverns
