--- domain/entity/Furnishing.ts
@@ -1,18 +1,0 @@
-import { Resources } from "./Resources";
-
-export class Tile {}
-
-export class Furnishing extends Tile {
-    constructor(id: FurnishingId, price: Resources) {
-        super();
-        this.id = id;
-        this.price = price;
-    }
-
-    id: FurnishingId;
-    price: Resources;
-}
-
-export enum FurnishingId {
-    DWELLING = "dwelling",
-}
