+++ domain/entity/index.ts
@@ -1,7 +1,7 @@
 export { ActionSpace, ActionSpaceId } from "./ActionSpace";
 export type { Action } from "./ActionSpace";
 
-export { Tile, Furnishing, FurnishingId } from "./Furnishing";
+export { Tile, Furnishing, FurnishingId } from "./Tile";
 
 export { ActionBoard, FurnishingBoard, Game } from "./Game";
